+++ networking/p2p/src/test/java/tech/pegasys/teku/networking/p2p/network/config/FilePrivateKeySourceTest.java
@@ -29,8 +29,7 @@ class FilePrivateKeySourceTest {
   @Test
   void shouldCreateKeyAndSaveToFile(@TempDir final Path tempDir) throws IOException {
     final Path file = tempDir.resolve("file.txt");
-    final PrivateKeySource privateKeySource =
-        new GetOrGenerateFilePrivateKeySource(file.toString());
+    final PrivateKeySource privateKeySource = new GeneratingFilePrivateKeySource(file.toString());
     final Bytes generatedBytes = privateKeySource.getPrivateKeyBytes();
     final Bytes savedBytes = Bytes.fromHexString(Files.readString(file));
 
@@ -42,8 +41,7 @@ class FilePrivateKeySourceTest {
     final Path file = tempDir.resolve("file.txt");
     final Bytes privateKey = Bytes.wrap(PrivateKeyGenerator.generate().bytes());
     Files.writeString(file, privateKey.toHexString());
-    final PrivateKeySource privateKeySource =
-        new GetOrGenerateFilePrivateKeySource(file.toString());
+    final PrivateKeySource privateKeySource = new GeneratingFilePrivateKeySource(file.toString());
 
     assertThat(privateKeySource.getPrivateKeyBytes()).isEqualTo(privateKey);
   }
@@ -53,8 +51,7 @@ class FilePrivateKeySourceTest {
     final Path file = tempDir.resolve("file.dat");
     final Bytes privateKey = Bytes.wrap(PrivateKeyGenerator.generate().bytes());
     Files.write(file, privateKey.toArray());
-    final PrivateKeySource privateKeySource =
-        new GetOrGenerateFilePrivateKeySource(file.toString());
+    final PrivateKeySource privateKeySource = new GeneratingFilePrivateKeySource(file.toString());
 
     assertThat(privateKeySource.getPrivateKeyBytes()).isEqualTo(privateKey);
   }
@@ -62,7 +59,7 @@ class FilePrivateKeySourceTest {
   @Test
   void shouldThrowExceptionIfInvalidFileName(@TempDir final Path tempDir) {
     final PrivateKeySource privateKeySource =
-        new GetOrGenerateFilePrivateKeySource(tempDir + "/invalid file name!!\0");
+        new GeneratingFilePrivateKeySource(tempDir + "/invalid file name!!\0");
     assertThatThrownBy(privateKeySource::getPrivateKeyBytes)
         .isInstanceOf(IllegalArgumentException.class)
         .hasMessageContaining("Not able to create or retrieve p2p private key file -");
@@ -71,7 +68,7 @@ class FilePrivateKeySourceTest {
   @Test
   void shouldThrowExceptionIfProvideDirectory(@TempDir final Path tempDir) {
     final PrivateKeySource privateKeySource =
-        new GetOrGenerateFilePrivateKeySource(tempDir.toString());
+        new GeneratingFilePrivateKeySource(tempDir.toString());
     assertThatThrownBy(privateKeySource::getPrivateKeyBytes)
         .isInstanceOf(RuntimeException.class)
         .hasMessageContaining("p2p private key file not found -");
